tlp-func-base, do_exit: use rm -f to remove files
_conf_tmp is a temporary file, not a directory,
so omit the more dangerous --recursive option.

The rm construct

   [ -f FILE ] && rm -f -- FILE

can also be written as

   rm -f -- FILE

since rm will ignore missing files in that case.
It will complain about missing permissions and directories, though.

To account for invalid call chains where _conf_tmp is not set,
add a "is _conf_tmp set?" check in front of the rm command.